+++ crates/ruma-html/CHANGELOG.md
@@ -1,5 +1,9 @@
 # [unreleased]
 
+Improvements:
+
+- Add support for deprecated HTML tags, according to Matrix 1.10
+
 # 0.1.0
 
 Initial release