+++ crates/ruma-html/CHANGELOG.md
@@ -1,5 +1,10 @@
 # [unreleased]
 
+Breaking Changes:
+
+- Do not export `Node` in the public API, it is not usable on its own and it is
+  not in the output of any public method.
+
 Improvements:
 
 - Add support for deprecated HTML tags, according to Matrix 1.10