+++ crates/ruma-html/CHANGELOG.md
@@ -8,6 +8,8 @@ Breaking Changes:
 Improvements:
 
 - Add support for deprecated HTML tags, according to Matrix 1.10
+- Allow to navigate through the HTML tree with `Html::first_child()`,
+  `Html::last_child()` or `Html::children()`
 
 # 0.1.0