+++ crates/ruma-html/CHANGELOG.md
@@ -4,6 +4,7 @@ Breaking Changes:
 
 - Do not export `Node` in the public API, it is not usable on its own and it is
   not in the output of any public method.
+- `Html::sanitize_with` now takes a reference to `SanitizerConfig`.
 
 Improvements: